fix layout of binary data preview in log view
